# swagger.yaml
swagger: '2.0'
info:
  description: >-
    
  version: 1.0.0
  title: Hackathon Scoring Project
host: localhost:8888
tags:
  - name: scores
  - name: users
schemes:
  - http
paths:
  /scores:
    get:
      tags:
        - scores
      summary: Returns all score
      operationId: getAllScore
      responses:
        '200':
          description: successful operation
        '500':
          description: Something went wrong
  /scores/{judgeId}:
    get:
      tags:
        - scores
      summary: Returns a list of score belong to a judgeID
      operationId: getScoreByJudge
      produces:
        - application/json
        - application/xml
      parameters:
        - name: "judgeId"
          in: "path"
          description: "ID of team to return (01~05)"
          required: true
          type: "string"
      responses:
        '200':
          description: successful operation
        '500':
          description: Something went wrong
    post:
      tags:
        - scores
      summary: Submit the score
      operationId: SubmittionWithJudge
      produces:
        - application/json
        - application/xml
      parameters:
        - name: "judgeId"
          in: "path"
          description: "ID of team to return (01~05)"
          required: true
          type: "string"
      responses:
        '200':
          description: successful operation
        '500':
          description: Something went wrong
  /scores/{judgeId}/{teamId}:
    put:
      tags:
        - scores
      summary: Save the score already filled
      description: ''
      operationId: scoreSave
      produces:
        - application/xml
        - application/json
      parameters:
        - name: "judgeId"
          in: "path"
          description: "ID of judge to update (01~05)"
          required: true
          type: "string"
        - name: "teamId"
          in: "path"
          description: "ID of team to update (A~K)"
          required: true
          type: "string"
        - in: body
          name: body
          description: score information
          required: true
          schema:
            $ref: '#/definitions/score'
      responses:
        '200':
          description: save temp success
        '500':
          description: Something went wrong
  /scores/avg:
    get:
      tags:
        - scores
      summary: Returns the average score of all
      operationId: getAvgScore
      responses:
        '200':
          description: successful operation
        '500':
          description: Something went wrong
  /users:
    get:
      tags:
        - users
      summary: Return all users
      operationId: getAllUsers
      responses:
        '200':
          description: successful operation
        '500':
          description: Something went wrong
    put:
      tags:
        - users
      summary: add a new user
      description: ''
      operationId: addUser
      produces:
        - application/json
        - application/xml
      parameters:
        - in: body
          name: body
          description: user object to be added into the system
          required: true
          schema:
            $ref: '#/definitions/user'
      responses:
        '200':
          description: login success
  /users/login:
    post:
      tags:
        - users
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      produces:
        - application/json
        - application/xml
      parameters:
        - in: body
          name: body
          description: User information
          required: true
          schema:
            $ref: '#/definitions/user'
      responses:
        '200':
          description: login success
        '400':
          description: Invalid username/password supplied
        '500':
          description: Something went wrong
  /users/logout:
    post:
      tags:
        - users
      summary: Logs out current logged in user session
      description: ''
      operationId: logoutUser
      produces:
        - application/json
        - application/xml
      parameters: []
      responses:
        default:
          description: successful operation
definitions:
  user:
    type: object
    properties:
      username:
        type: string
        format: string
      password:
        type: string
    xml:
      name: user
  score:
    type: object
    properties:
      teamId:
        type: "string"
        enum:
         - "A"
         - "B"
         - "C"
         - "D"
         - "E"
         - "F"
         - "G"
         - "H"
         - "I"
         - "J"
         - "K"
      rank:
        type: "integer"
        format: "int16"
      name:
        type: "string"
      topic:
        type: "string"
      maintain:
        type: "integer"
        format: "int16"
      innov:
        type: "integer"
        format: "int16"
      design:
        type: "integer"
        format: "int16"
      skill:
        type: "integer"
        format: "int16"
      demo:
        type: "integer"
        format: "int16"
      result:
        type: "number"
        format: "double"
        description: Total score
      comment:
        type: "string"
      judgeId:
        type: "string"
        enum:
         - "01"
         - "02"
         - "03"
         - "04"
         - "05"
      complete:
        type: "string"
        description: Submitted or not
        enum:
         - "false"
         - "true"
    xml:
      name: Category
  avg:
    type: object
    properties:
      teamId:
        type: "string"
        enum:
         - "A"
         - "B"
         - "C"
         - "D"
         - "E"
         - "F"
         - "G"
         - "H"
         - "I"
         - "J"
         - "K"
      rank:
        type: "integer"
        format: "int16"
      name:
        type: "string"
      topic:
        type: "string"
      maintain:
        type: "integer"
        format: "int16"
      innov:
        type: "integer"
        format: "int16"
      design:
        type: "integer"
        format: "int16"
      skill:
        type: "integer"
        format: "int16"
      demo:
        type: "integer"
        format: "int16"
      result:
        type: "number"
        format: "double"
        description: Total score
      comment:
        type: "string"
    xml:
      name: Category